requirements.txt
# Data ingestion (yfinance, Finnhub)
pandas>=1.5.0
numpy>=1.22
yfinance>=0.2.0
requests>=2.28.0
finnhub-python>=1.0.0

# FastAPI + Swagger
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Scheduling
apscheduler>=3.10.0

# Database (PostgreSQL / Supabase + pgvector)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
python-dotenv>=1.0.0

# Embeddings for skill graph indexer
openai>=1.0.0
